+++ packages/client/src/entities/lobby/models/index.ts
@@ -8,4 +8,15 @@ export interface Player {
   role: string
 }
 
+export type Message = {
+  chat_id: number
+  content: string
+  file: string
+  id: number
+  is_read: boolean
+  time: string
+  type: 'message'
+  user_id: number
+}
+
 export type { LobbyView } from './lobby-view'
